+++ DelvUI/Interface/Party/PartyFramesBar.cs
@@ -7,7 +7,7 @@ using ImGuiNET;
 using System;
 using System.Globalization;
 using System.Numerics;
-using Dalamud.Logging;
+using Dalamud.Game.ClientState.Objects.Types;
 
 namespace DelvUI.Interface.Party
 {
@@ -161,7 +161,7 @@ namespace DelvUI.Interface.Party
 
             // click
             bool isHovering = ImGui.IsMouseHoveringRect(Position, Position + _config.Size);
-            var character = Member.Character;
+            Character? character = Member.Character;
 
             if (isHovering)
             {
@@ -195,15 +195,25 @@ namespace DelvUI.Interface.Party
             }
 
             // bg
-            PluginConfigColor bgColor = Member.RaiseTime != null && _raiseTrackerConfig.Enabled && _raiseTrackerConfig.ChangeBackgroundColorWhenRaised
-                ? _raiseTrackerConfig.BackgroundColor
-                : Member.InvulnStatus?.InvulnTime != null && _invulnTrackerConfig.Enabled && _invulnTrackerConfig.ChangeBackgroundColorWhenInvuln
-                    ? Member.InvulnStatus?.InvulnId == 811
-                        ? _invulnTrackerConfig.WalkingDeadBackgroundColor
-                        : _invulnTrackerConfig.BackgroundColor
-                    : _config.ColorsConfig.UseDeathIndicatorBackgroundColor && Member.HP <= 0
-                        ? _config.ColorsConfig.DeathIndicatorBackgroundColor
-                        : _config.ColorsConfig.BackgroundColor;
+            PluginConfigColor bgColor;
+            if (Member.RaiseTime != null && _raiseTrackerConfig.Enabled && _raiseTrackerConfig.ChangeBackgroundColorWhenRaised)
+            {
+                bgColor = _raiseTrackerConfig.BackgroundColor;
+            }
+            else if (Member.InvulnStatus?.InvulnTime != null && _invulnTrackerConfig.Enabled && _invulnTrackerConfig.ChangeBackgroundColorWhenInvuln)
+            {
+                bgColor = Member.InvulnStatus?.InvulnId == 811 ? _invulnTrackerConfig.WalkingDeadBackgroundColor : _invulnTrackerConfig.BackgroundColor;
+            }
+            else if (_config.ColorsConfig.UseDeathIndicatorBackgroundColor && Member.HP <= 0)
+            {
+                bgColor = _config.RangeConfig.Enabled
+                    ? GetDistance(character, _config.ColorsConfig.DeathIndicatorBackgroundColor)
+                    : _config.ColorsConfig.DeathIndicatorBackgroundColor;
+            }
+            else
+            {
+                bgColor = _config.ColorsConfig.BackgroundColor;
+            }
 
             drawList.AddRectFilled(Position, Position + _config.Size, bgColor.Base);
 
@@ -220,12 +230,9 @@ namespace DelvUI.Interface.Party
             var hpFillSize = new Vector2(_config.Size.X * hpScale, _config.Size.Y);
             PluginConfigColor? hpColor = GetColor(hpScale);
 
-            var distance = character != null ? character.YalmDistanceX : byte.MaxValue;
             if (_config.RangeConfig.Enabled)
             {
-                var currentAlpha = hpColor.Vector.W * 100f;
-                var alpha = _config.RangeConfig.AlphaForDistance(distance, currentAlpha) / 100f;
-                hpColor = new(hpColor.Vector.WithNewAlpha(alpha));
+                hpColor = GetDistance(character, hpColor);
             }
 
             DrawHelper.DrawGradientFilledRect(Position, hpFillSize, hpColor, drawList);
@@ -250,7 +257,7 @@ namespace DelvUI.Interface.Party
             // border
             var borderPos = Position - Vector2.One;
             var borderSize = _config.Size + Vector2.One * 2;
-            var color = borderColor != null ? borderColor.Base : _config.ColorsConfig.BorderColor.Base;
+            var color = borderColor?.Base ?? _config.ColorsConfig.BorderColor.Base;
             drawList.AddRect(borderPos, borderPos + borderSize, color);
 
             // role/job icon
@@ -310,6 +317,15 @@ namespace DelvUI.Interface.Party
             {
                 drawList.AddRectFilled(Position, Position + _config.Size, _config.ColorsConfig.HighlightColor.Base);
             }
+        }
+
+        private PluginConfigColor GetDistance(Character? character, PluginConfigColor color)
+        {
+            byte distance = character != null ? character.YalmDistanceX : byte.MaxValue;
+            float currentAlpha = color.Vector.W * 100f;
+            float alpha = _config.RangeConfig.AlphaForDistance(distance, currentAlpha) / 100f;
+
+            return new PluginConfigColor(color.Vector.WithNewAlpha(alpha));
         }
 
         // need to separate elements that have their own window so clipping doesn't get messy