+++ DelvUI/Interface/Party/PartyFramesCleanseTracker.cs
@@ -8,13 +8,34 @@ using DelvUI.Helpers;
 
 namespace DelvUI.Interface.Party
 {
-    public class PartyFramesCleanseTracker
+    public class PartyFramesCleanseTracker : IDisposable
     {
         private PartyFramesCleanseTrackerConfig _config;
         public PartyFramesCleanseTracker()
         {
             _config = ConfigurationManager.Instance.GetConfigObject<PartyFramesCleanseTrackerConfig>();
             ConfigurationManager.Instance.ResetEvent += OnConfigReset;
+        }
+
+        ~PartyFramesCleanseTracker()
+        {
+            Dispose(false);
+        }
+
+        public void Dispose()
+        {
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+        protected void Dispose(bool disposing)
+        {
+            if (!disposing)
+            {
+                return;
+            }
+
+            ConfigurationManager.Instance.ResetEvent -= OnConfigReset;
         }
 
         public void OnConfigReset(ConfigurationManager sender)
@@ -31,11 +52,7 @@ namespace DelvUI.Interface.Party
 
             foreach (var member in partyMembers)
             {
-                if (member.Character == null || member.ObjectId == 0)
-                {
-                    member.HasDispellableDebuff = false;
-                    continue;
-                }
+                member.HasDispellableDebuff = false;
 
                 if (member.Character is not BattleChara battleChara)
                 {